backend/requirements.txt
Flask
pandas
requests
sentence-transformers
google-generativeai
python-dotenv
scikit-learn
